moved the library support in a sub module.
